(Don't break) Hudfunctions, lobby player, spectator, and car
Major changes in the scripts. The main script that acts as a network
manager is HudFunctions. Added a lobby player, spectator player, and the
car now spawns. Still needs fixing, but there is now a hud to spawn the
player as each. There is also a start menu now that allows users to
create a host or to connect to one.